+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2022-12-04
+
+### Added
+- Support for providing menu items as a prepopulated array via the menu_items option.
+- Support for callables as templates, enabling template string to be defined dynamically.
+
 ## [0.10.0] - 2021-09-02
 
 ### Added